image_classification/pyproject.toml
[tool.poetry]
name = "image-classification"
version = "0.1.0"
description = "Image Classification API using TensorFlow"
authors = ["Aurora <aolaya@berkeley.edu>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"  # Updated Python version constraint
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
python-multipart = "^0.0.6"
pillow = "^10.1.0"
numpy = "^1.26.0"
tensorflow-cpu = "^2.14.0"
tensorflow-io-gcs-filesystem = "0.31.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"